Refactor: Move K8s helper functions to benji-k8s-tools package
