+++ images/benji-k8s/k8s-tools/setup.py
@@ -12,9 +12,7 @@ setup(name='benji-k8s-tools',
       package_dir={
           '': 'src',
       },
-      install_requires=[
-          'benji',
-      ],
+      install_requires=['benji', 'kubernetes>=10.0.0,<11'],
       entry_points="""
         [console_scripts]
             benji-backup-pvc = benji.k8s_tools.scripts.backup_pvc:main
